+++ src/rxcore/models/IMagnifier.ts
@@ -1,6 +1,8 @@
 export interface IMagnifier {
     setContext(ctx: CanvasRenderingContext2D, width: number, height: number): void;
+    setCanvas(baseCanvas: HTMLCanvasElement, overlayCanvas : HTMLCanvasElement, snapCanvas : HTMLCanvasElement): void;
     setOverlayContext(ctx: CanvasRenderingContext2D): void;
+    setSnapContext(ctx: CanvasRenderingContext2D): void;
     setActive(state: boolean): void;
     toggle(): void;
     resetContext(): void;
@@ -10,5 +12,14 @@ export interface IMagnifier {
     drawOverlayPoint(x: number, y: number, color?: string, size?: number): void;
     drawCrosshair(color?: string, size?: number, box?: boolean): void;
     update(mousePos: { x: number; y: number }): Promise<void>;
+    suspend(state: boolean): void;
+    setUseMagnifyAlign(enable : boolean) : void;
+    enterPrecisionMode(anchor: number) : void;
+    exitPrecisionMode() : void;
+    precisionMode : boolean;
+    onEnterPrecision() : void;
+    onExitPrecision() : void;
+    
+
   }